refactor(dashboard): the deploy flow gets a folder of its own
components/apps held forty-eight files, sixteen of which were the deploy
form and the two shells that open it. They sit beside device/, files/,
login/ and logs/ now, where a feature folder is what the rest of the
components directory already is.
