+++ apps/dashboard/src/components/apps/app-environment-dialog-content.tsx
@@ -11,9 +11,9 @@ import { Field, FieldError } from '@repo/ui/components/field';
 import { DialogBody } from '@repo/ui/custom/dialog-body';
 import { PencilIcon } from 'lucide-react';
 import { useState } from 'react';
-import { DeployDoneButton } from '#components/apps/deploy-done-button.tsx';
-import { DeployProgress } from '#components/apps/deploy-progress.tsx';
 import { EnvironmentTable } from '#components/apps/environment-table.tsx';
+import { DeployDoneButton } from '#components/deploy/deploy-done-button.tsx';
+import { DeployProgress } from '#components/deploy/deploy-progress.tsx';
 import { greyedReason } from '#lib/app-actions.ts';
 import { useAppActions } from '#lib/hooks/use-app-actions.ts';
 import { useDeployRun } from '#lib/hooks/use-deploy-run.ts';
