+++ apps/dashboard/src/components/deploy/binary-drop-zone.tsx
